FragmentosEnDetallesDeLasFrutas

+++ app/src/main/java/com/example/frutapp/data/model/NutritionsDTO.kt
@@ -4,6 +4,6 @@ data class NutritionsDTO(
   val carbohydrates: Double,
   val protein: Double,
   val fat: Double,
-  val calories: Int,
+  val calories: Double,
   val sugar: Double
 )
